SNA4Slack_API/requirements.txt
gunicorn
cassandra-driver
Flask
flask_cassandra
flasgger
flask_restful
selenium
pyvirtualdisplay
networkx
dateutils
pymongo
findspark
re
